requirements.txt
Django==5.2.3
djangorestframework==3.16.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
sentence-transformers==5.0.0
pgvector==0.4.1
huggingface-hub==0.33.4
